# src/schemas/openapi_v2_3.yaml
openapi: 3.1.0
info:
  title: PACT API Specification
  summary: Technical Specifications for PCF Data Exchange
  version: 2.3.1
  contact: 
    name: WBCSD PACT Project
    url: https://github.com/wbcsd
    email: pact@wbcsd.org
  description: |
    This document describes the technical specifications for the exchange of 
    Product Carbon Footprint (PCF) data between systems. The specification is 
    maintained by the WBCSD PACT project and intended to be used by organizations 
    that wish to exchange PCF data with other organizations. The specifications 
    are designed to be technology-agnostic.
...
servers: 
- url: https://example.com
paths:
...
  /2/events:
    post:
      summary: Action Event
      description: |
        Post an event to the system. The event will be processed by the system 
        and may trigger further actions.
      operationId: post_event
      tags: [event]
      requestBody:
        content:
          application/cloudevents+json:
            schema:
              oneOf:
              - $ref: '#/components/schemas/RequestCreatedEvent'
              - $ref: '#/components/schemas/RequestFulfilledEvent'
              - $ref: '#/components/schemas/RequestRejectedEvent'
              - $ref: '#/components/schemas/PublishedEvent'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties: 
                  code: 
                    type: string
                  message: 
                    type: string
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        '403':
          $ref: '#/components/responses/ForbiddenResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
      security:
      - BearerAuth: []
components:
  schemas:
    ProductFootprint:
      description: |
        `ProductFootprint` is a data type which represents the carbon footprint
        of a product under a specific scope ([[#dt-carbonfootprint-scope]])
        and with values calculated in accordance with the [=PACT Methodology=].

        The objective of a `ProductFootprint` is to provide interoperability between
        the creator (the [=data owner=]) and the consumer (the [=data recipient=]) of
        ProductFootprints. The details on the exchange of ProductFootprints are
        specified in [[#api]].

        Conceptually, the data type <{ProductFootprint}> is modeled as a multi-purpose
        container for product-specific emissions factors which is supported by
        extensibility through [=Data Model Extensions=].

        Data Model Extensions enable [=data owners=] to exchange additional information
        related to a product with [=data recipients=]. The details are specified
        in [[#datamodelextension]] as well as [[!EXTENSIONS-GUIDANCE]], and [[!DATA-MODEL-EXTENSIONS]].

        Each `ProductFootprint` can and should be updated over time, for instance to
        incorporate new or refined data from [=data owners=] (see [[#lifecycle]]).
      type: object
      title: ProductFootprint
      required:
      - id
      - specVersion
      - version
      - created
      - status
      - comment
      - companyName
      - companyIds
      - productDescription
      - productIds
      - productCategoryCpc
      - productNameCompany
      - pcf
      properties:
        id:
          description: |
            A unique identifier that a system uses to refer to the entire dataset of the PCF. 
            This is typically an automatically-generated number by the technology  solution to maintain 
            the required technical references to data within the system.
            UUID v4 [RFC4122](https://tools.ietf.org/html/rfc4122).
          type: string
          format: uuid
          examples:
          - "f4b1225a-bd44-4c8e-861d-079e4e1dfd69"
        specVersion:
          description: |
            The version of the PACT Tech Specs that the data being shared complies with. 
            This is a string in the format of "major.minor.patch" (e.g. "3.0.0").
          type: string
          pattern: '^\d+\.\d+\.\d+(-\d{8})?$'
          examples:
          - "3.0.0"
        precedingPfIds:
          description: |
            A given PCF may change over time, due to updates to the calculation. 
            This field lists all id's that reflect "past versions" of the current PCF. 
            This is typically an automatically generated field of unique identifiers, 
            maintained by the technology solution.
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          uniqueItems: true
          examples:
          - ["f4b1225a-bd44-4c8e-861d-079e4e1dfd69", 
             "079e425a-464f-528d-341d-4a944a1dfd70"]
        version:
          description: |
            The version number of the PCF; if changes are made to the PCF, 
            this version number will increment.
          type: integer
          format: int32
          minimum: 0
          examples:
          - 1
          - 3303
        created:
          description: |
            The date and time when the PCF was created. This is typically an automatically 
            generated field by the software solution.
          type: string
          format: date-time
          examples:
          - '2024-10-31T00:00:00Z'
        updated:
          description: |
            The date and time when the PCF was last updated. This is typically an automatically 
            generated field by the software solution. If the data has never been updated, 
            do not include this property.
          type: string
          format: date-time
        status:
          description: |
            The status of the PCF. `Active` means that the PCF is the most recent version 
            and is the one that can be used CAN be used by a data recipients, e.g. for product 
            footprint calculations. `Deprecated` means that the PCF is no longer 
            the most recent version and SHOULD NOT be used by data recipients.
            
            See [[#lifecycle]] for details.
          type: string
          enum:
            - Active
            - Deprecated
          examples: 
          - Active
        statusComment:
          type: string
          description: |
            TODO: Move this description to #lifecycle

            If defined, a descriptive reasoning explaining the current status of 
            the PCF, what was changed since the last version, etc. If the PCF was 
            changed since a previous version, indicate all methodological and/or 
            production process change(s) that occurred to result in the PCF change. 
            For example, include the relevant change(s) from the list below:

            Methodological:
            - Access to new Emission Factor data (database, supplier-specific, company-specific)
            - Updated upstream data (i.e. upstream supplier updated their PCF based on methodology change)

            Production Process:
            - Change in process
            - Change in feedstock
            - Change from conventional to certified sustainable material
            - Change in energy source
            - Change in upstream supplier
            - Updated upstream data (i.e. upstream supplier updated their PCF based on process change)
          
            See [[#lifecycle]] for details.
        validityPeriodStart:
          description: |
            TODO: Create a separate paragraph on validity period and reference period

            The <dfn>validity period</dfn> is the time interval during which the 
            ProductFootprint is declared as valid for use by a receiving [=data recipient=].

            The validity period is OPTIONAL defined by the properties 
            <{ProductFootprint/validityPeriodStart}> (including) and 
            <{ProductFootprint/validityPeriodEnd}> (excluding).

            If no validity period is specified, the ProductFootprint is valid for 
            3 years starting with <{CarbonFootprint/referencePeriodEnd}>.

            If a validity period is to be specified, then
              1. the value of <{ProductFootprint/validityPeriodStart}> MUST be defined with value 
                 greater than or equal to the value of <{CarbonFootprint/referencePeriodEnd}>.
              2. the value of <{ProductFootprint/validityPeriodEnd}> MUST be defined with value

                 1. strictly greater than <{ProductFootprint/validityPeriodStart}>, and
                 2. less than or equal to <{CarbonFootprint/referencePeriodEnd}> + 3 years.

          summary: |
            The start date and time of the validity period. The period of time between the 
            validityPeriodStart and validityPeriodEnd is defined as the "validity period", 
            and represents the time period during which the PCF is valid. Specifying the 
            validity period is optional. If this is not specified, then it is assumed the 
            PCF is valid for 3 years, beginning with the referencePeriodEnd. 
          type: string
          format: date-time
        validityPeriodEnd:
          description: |
            The end date and time of the validity period. See validityPeriodStart for more details.
          type: string
          format: date-time
        companyName:
          description: |
            The name of the company that is the ProductFootprint Data Owner
          $ref: '#/components/schemas/NonEmptyString'
        companyIds:
          description: |
            The non-empty set of Uniform Resource Names ([[!RFC8141|URN]]). Each value of 
            this set is supposed to uniquely identify the ProductFootprint Data Owner. 
            TODO: See Company Identifier section for details
          summary: |
            A list of company identifiers which represent the companies that are considered 
            data owners of the PCF. In a large organization, this may include multiple
            subsidiaries or legal entities of a given corporation, hence the need to provide 
            multiple identifiers. A list of only one identifier is also valid. The format 
            of each companyID MUST be provided as a Uniform Resource Names ([[!RFC8141|URN]]),
            which helps ensure the data provided is standardized and can be interpreted by the 
            receiving system.
          examples:
          - ["urn:company:example:company1"]
          - ["urn:company:example:company1","urn:company:example:company2"]
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Urn'
        productDescription:
          description: |
            The free-form description of the product, including any additional relevant information 
            such as production technology, packaging, process, feedstock and technical parameters 
            (e.g. dimensions). Products which are services (i.e. consulting) should include a short
            description of the service.
          summary: |
            A description of the product, including any additional relevant information 
            such as production technology, production technology, packaging, process, feedstock 
            and technical parameters (e.g. dimensions). Products which are services 
            (i.e. consulting) should include a short description of the service.
          type: string
        productIds:
          description: |
            The non-empty set of [=ProductIds=] in [=URN=] format. Each of the values in the set is 
            supposed to uniquely identify the product. See {#product-identifiers} for syntax and 
            examples.
          summary: |
            All relevant product identifiers to identify the product for which the provided 
            PCF was calculated (e.g. supplier part number, GTIN, article number, CPC 
            classification, CAS Number, BOM Version, batch number, etc.)
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Urn'
        productClassifications:
          description: |
            The non-empty set of [=ProductClassifications=] in [=URN=] format. Each of the values in the set can classify the product as part of distinct groupings and categorizations. See {#product-identifiers}. 
            
            Advisement: Supersedes productCategoryCpc deprecated in version 2.3.
          summary: |
            A list of classification or category identifiers in URN format. Use well known urn's here, 
            or adhere to recommended urn:pact: format. This replaces the productCategoryCpc property.
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Urn'
        productCategoryCpc:
          description: |
            Advisement: DEPRECATED in 3.0. Superseded by <{ProductFootprint/productClassifications}>.

            The UN Central Product Classification (CPC) that the given product belongs to.
          summary: |
            The UN Central Product Classification Code (CPC) for the given product. The relevant 
            CPC code may be identified here https://unstats.un.org/unsd/classifications/Econ/cpc 
            Advisement: This property will become optional in v3
          $ref: '#/components/schemas/NonEmptyString'
        productNameCompany:
          description: |
            The non-empty trade name of the product.
          summary: |
            The trade name of the product in order for it to be recognizable by the receiver of the PCF information          
          $ref: '#/components/schemas/NonEmptyString'
        comment:
          description: |
            Advisement: OPTIONAL in 3.0.

            Additional information related to the product footprint.

            Whereas the property <{ProductFootprint/productDescription}> contains product-level information, <{ProductFootprint/comment}> SHOULD be used for information and instructions related to the calculation of the footprint, or other information which informs the ability to interpret, to audit or to verify the Product Footprint.
          summary: |
            Any additional information related to the PCF. Whereas the property productDescription 
            contains product-level information, comment should be used for information and instructions 
            related to the calculation of the PCF, or other information which informs the ability to interpret
            (e.g. LUC not included as unable to calculate LUC), to audit or to verify the PCF          
            Advisement: This property will become OPTIONAL in version 3 of the Technical Specifications.
          type: string
        pcf:
          description: |
            The carbon footprint of the given product with value conforming to the data 
            type <{CarbonFootprint}>.
          $ref: '#/components/schemas/CarbonFootprint'
        extensions: 
          type: array
          description: |
            If defined, 1 or more data model extensions associated with the ProductFootprint.
            See <{DataModelExtension}> for details.
          items: 
            $ref: '#/components/schemas/DataModelExtension'
...
    DataModelExtension:
      title: DataModelExtension
      description: |
        Each data model extension MUST be a valid JSON object conforming with the
        [JSON Representation of a Data Model Extension](https://wbcsd.github.io/data-model-extensions/spec/#instantiation).

        See [[!DATA-MODEL-EXTENSIONS]] for technical details and [[!EXTENSIONS-GUIDANCE]] for data model extension guidance.
      example: |
        Example imaginary Data Model Extension for encoding shipment-related data, encoded in JSON:
        ```json
        {
          "specVersion": "2.0.0",
          "dataSchema": "https://catalog.carbon-transparency.com/shipment/1.0.0/data-model.json",
          "data": {
            "shipmentId": "S1234567890",
            "consignmentId": "Cabc.def-ghi",
            "shipmentType": "PICKUP",
            "weight": 10,
            "transportChainElementId": "ABCDEFGHI"
          }
        }
        ```
      type: object
      required:
      - specVersion
      - dataSchema
      - data
      properties:
        specversion:
          type: string
          description: |
            The version of the Data Model Extension specification. The value 
            MUST be a string in the format major.minor.patch as defined in 
            Semantic Versioning 2.0.0.
        dataSchema:
          type: string
          format: uri
          description: |
            The value MUST be the URL to the publicly accessible Extension Schema File
        documentation:
          type: string
          format: uri
          description: |
            The value MUST be the URL to the publicly accessible Extension Documentation.
        data:
          type: object
          description: |
            The value MUST be a JSON Object that conforms to the extension schema file 
            referenced by the dataSchema property.
...
    NonEmptyString:
      type: string
      minLength: 1
    Urn:
      type: string
      format: urn
      pattern: '^([uU][rR][nN]):'
...
    PublishedEvent:
      title: ProductFootprint Published
      description: Notification that a ProductFootprint has been published (either new or updated).
      type: object
      required:
      - type
      - specversion
      - id
      - source
      - time
      - data
      properties:
        type:
          type: string
          const: org.wbcsd.pathfinder.ProductFootprint.Published.v1
          description: Event type identifier.
        specversion:
          type: string
          description: CloudEvents version.
        id:
          type: string
          description: Event identifier. Must be able to uniquely identify the event by source and id.
        source:
          type: string
          format: uri
          description: The domain and endpoint of the application from which the event originates.
        time:
          type: string
          format: date-time
          description: The time the event occurred.
        data:
          type: object
          description: data object
          properties:
            pfIds:
              type: array
              description: The list of Product Footprint Ids that have been published or updated by the data owner.
              items:
                type: string
                format: uuid
    RequestCreatedEvent:
      title: ProductFootprint Request Created
      description: A data recipient request the data owner a ProductFootprint.
      type: object
      required:
      - type
      - specversion
      - id
      - source
      - time
      - data
      properties:
        type:
          type: string
          const: org.wbcsd.pathfinder.ProductFootprintRequest.Created.v1
          description: Event type identifier.
        specversion:
          type: string
          description: CloudEvents version.
        id:
          type: string
          description: Event identifier. Must be able to uniquely identify the event by source and id.
        source:
          type: string
          format: uri
          description: The domain and endpoint of the application from which the event originates.
        time:
          type: string
          format: date-time
          description: The time the event occurred.
        data:
          type: object
          description: data object
          required:
          - pf
          properties:
            pf:
              type: object
              description: ProductFootprintFragment specifying the requested footprint(s).
            comment:
              type: string
    RequestFulfilledEvent:
      title: ProductFootprint Request Fulfilled
      description: Notification that a ProductFootprint request has been fulfilled.
      type: object
      properties:
        type:
          type: string
          const: org.wbcsd.pathfinder.ProductFootprintRequest.Fulfilled.v1
          description: Event type identifier.
        specversion:
          type: string
          description: CloudEvents version.
        id:
          type: string
          description: Event identifier. Must be able to uniquely identify the event by source and id.
        source:
          type: string
          format: uri
          description: The domain and endpoint of the application from which the event originates.
        time:
          type: string
          format: date-time
          description: The time the event occurred.
        data:
          type: object
          description: data object
          required:
            - requestEventId
            - pfs
          properties:
            requestEventId:
              type: string
            pfs:
              type: array
              items:
                $ref: '#/components/schemas/ProductFootprint'
    RequestRejectedEvent:
      title: ProductFootprint Request Rejected
      description: Notification that a request for a ProductFootprint can not be fulfilled.
      type: object
      properties:
        type:
          type: string
          const: org.wbcsd.pathfinder.ProductFootprintRequest.Rejected.v1
          description: Event type identifier.
        specversion:
          type: string
          description: CloudEvents version.
        id:
          type: string
          description: Event identifier. Must be able to uniquely identify the event by source and id.
        source:
          type: string
          format: uri
          description: The domain and endpoint of the application from which the event originates.
        time:
          type: string
          format: date-time
          description: The time the event occurred.
        data:
          type: object
          description: data object
          required:
            - requestEventId
            - error
          properties:
            requestEventId:
              type: string
            error:
              $ref: '#/components/schemas/Error'
    # ---------------
    # Error Schema
    Error:
      title: Error
      description: Object with error code and descripton. See Chapter "Error
        Codes" of the Tech Specs for more details.
      type: object
      properties:
        code:
          type: string
          description: Error code identifier.
          enum:
            - BadRequest # 400
            - AccessDenied # 403
            - TokenExpired # 401
            - NotFound # 404 NoSuchFootprint
            - InternalError # 500
            - NotImplemented #501
        message:
          type: string
          description: Human readable error message.
      required:
      - code
      - message
  securitySchemes:
    BearerAuth:
      description: OAuth2 Client Credentials Grant (RFC6749 4.4)
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: '/auth/token'
          scopes:
            footprint:list: Ability to list footprints
            footprint:read: Ability to access individual footprints
  responses:
    BadRequestResponse:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundResponse:
      description: Not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedResponse:
      description: Unauthorized.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ForbiddenResponse:
      description: Access Denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalErrorResponse:
      description: Internal Error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'